Fix: 원화 데이터 BigDecimal -> Long

+++ src/main/java/com/example/fisafroexpay/dto/TransferRequest.java
@@ -14,5 +14,5 @@ public class TransferRequest {
     private String accountNumber;  // 보내는 사람 계좌번호
     private String receiverAccountNumber;  // 받는 사람 계좌번호
     private String receiverCurrencyCode;  // 받는 사람 통화 코드
-    private BigDecimal amount;  // 송금 금액
+    private Long amount;  // 송금 금액
 }
